pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rostering-app"
version = "1.0.0"
description = "A Flask-based rostering application for staff management"
readme = "readme.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["flask", "rostering", "staff-management", "web-app"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Flask",
]
requires-python = ">=3.9"
dependencies = [
    "Flask==2.3.3",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Migrate==3.1.0",
    "Flask-Reuploaded==1.2.0",
    "Flask-Cors==3.0.10",
    "Flask-JWT-Extended==4.4.4",
    "Flask-Admin==1.6.1",
    "Werkzeug>=3.0.0",
    "click==8.1.3",
    "gunicorn==20.1.0",
    "gevent==23.9.1",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.1",
    "rich==13.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.0.1",
    "coverage",
]

[project.scripts]
rostering-app = "wsgi:app"

[project.urls]
Homepage = "https://github.com/yourusername/rostering-app"
Repository = "https://github.com/yourusername/rostering-app.git"
Issues = "https://github.com/yourusername/rostering-app/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["App*"]

[tool.setuptools.package-data]
"App" = ["templates/*", "templates/**/*", "static/*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["App/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"